assembler: refactor address of machine instruction to be uint
Making the machine instruction a uint means we do the parsing when we generate the machine instruction from objdump, etc. 
It also means that we have less errors we have to check later on, i.e. if the uint parsing fails. The address should always be a number, so storing it as a string isn't useful.
Also updating the tests for this.