assembler: support arm64 supported opcode translation
Adding support for ARM64 to translate supported opcodes into Go syntax. Note that this is currently disabled inside of asm2go, because the 1.10.2 assembler doesn't understand the disassembler's output syntax from tip. 
Also adding test for arm with forced translation of a supported opcode, and tests for ARM64 supported opcodes (both with translation forced and not forced)